Code review remarks: Move hashPacket and balancePacket
